+++ example/spawn_process.dart
@@ -4,19 +4,29 @@
 
 import 'dart:async';
 
-// TODO: Change to io.dart once this feature is published.
-import 'package:io/src/process_manager.dart';
+import 'package:io/io.dart';
 
-/// Runs `dartfmt` and `pub` via [ProcessManager.spawn] as an example.
+/// Runs `dartfmt` commands and `pub publish`.
 Future<Null> main() async {
   final manager = new ProcessManager();
+
+  // Runs dartfmt --version and outputs the result via stdout.
   print('Running dartfmt --version');
   var spawn = await manager.spawn('dartfmt', arguments: ['--version']);
   await spawn.exitCode;
+
+  // Runs dartfmt -n . and outputs the result via stdout.
   print('Running dartfmt -n .');
   spawn = await manager.spawn('dartfmt', arguments: ['-n', '.']);
   await spawn.exitCode;
+
+  // Runs pub publish. Upon hitting a blocking stdin state, you may directly
+  // output to the processes's stdin via your own, similar to how a bash or
+  // shell script would spawn a process.
   print('Running pub publish');
-  spawn = await manager.spawn('pub', arguments: ['pub', 'publish']);
-  await ProcessManager.terminateStdIn();
+  spawn = await manager.spawn('pub', arguments: ['publish']);
+  await spawn.exitCode;
+
+  // Closes stdin for the entire program.
+  await sharedStdIn.terminate();
 }